# reservation/docs/swagger-ui/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  APIError:
    properties:
      error:
        description: |-
          The status error message.
          in: body
        type: string
        x-go-name: ErrMsg
      message:
        description: |-
          The error message.
          in: body
        type: string
        x-go-name: Message
      status:
        description: |-
          The HTTP status code.
          in: body
        format: int64
        type: integer
        x-go-name: Status
      timestamp:
        description: |-
          The timestamp of the error.
          in: body
        type: string
        x-go-name: Timestamp
    required:
    - timestamp
    - status
    - error
    - message
    title: APIError is the error response for the API.
    type: object
    x-go-package: github.com/MortalHappiness/VaccineReservationSystem/go-utils/apierrors
  HospitalModel:
    description: Hospital model is the body format of HospitalResponse
    properties:
      address:
        description: The hospital address
        example: No.1, Sec. 4, Roosevelt Road, Taipei, 10617 Taiwan
        type: string
        x-go-name: Address
      county:
        description: The hospital County
        example: New Taipei City
        type: string
        x-go-name: County
      id:
        description: The hospital id
        example: "0001"
        type: string
        x-go-name: ID
      name:
        description: The hospital name
        example: Taipei City Hospital Heping Fuyou Branch
        type: string
        x-go-name: Name
      township:
        description: The hospital Township
        example: Banqiao
        type: string
        x-go-name: Township
      vaccineCnt:
        additionalProperties:
          format: int64
          type: integer
        description: The hospital vaccines
        example:
          AZ: 100
          BNT: 200
        type: object
        x-go-name: VaccineCnt
    required:
    - id
    - name
    - county
    - township
    - address
    - vaccineCnt
    type: object
    x-go-package: github.com/MortalHappiness/VaccineReservationSystem/go-utils/models
  ReservationModel:
    description: ReservationModel is the body format of ReservationResponse
    properties:
      completed:
        description: The vaccination is completed
        example: false
        type: boolean
        x-go-name: Completed
      date:
        description: The reservation date
        example: 1653974953
        format: int64
        type: integer
        x-go-name: Date
      hospital:
        $ref: '#/definitions/HospitalModel'
      id:
        description: The reservation id
        example: "0001"
        type: string
        x-go-name: ID
      user:
        $ref: '#/definitions/UserModel'
      vaccinetype:
        description: The reservation vaccinetype
        example: BNT
        type: string
        x-go-name: VaccineType
    required:
    - id
    - user
    - hospital
    - vaccinetype
    - date
    - completed
    type: object
    x-go-package: github.com/MortalHappiness/VaccineReservationSystem/go-utils/models
  UserModel:
    description: UserModel is the body format of UserResponse
    properties:
      address:
        description: The user address
        example: No.1, Sec. 4, Roosevelt Road, Taipei, 10617 Taiwan
        type: string
        x-go-name: Address
      birthDay:
        description: The user birthday
        example: 2022/05/23
        type: string
        x-go-name: BirthDay
      gender:
        description: The user gender
        example: male
        type: string
        x-go-name: Gender
      healthCardID:
        description: The user healthCardID
        example: "000011112222"
        type: string
        x-go-name: HealthCardID
      name:
        description: The user name
        example: bob
        type: string
        x-go-name: Name
      nationID:
        description: The user nation ID
        example: A123456789
        type: string
        x-go-name: NationID
      phone:
        description: The user phone number
        example: "0912345678"
        type: string
        x-go-name: Phone
      vaccines:
        description: The user inoculated vaccines
        example:
        - AZ
        - BNT
        items:
          type: string
        type: array
        x-go-name: Vaccines
    required:
    - nationID
    - healthCardID
    type: object
    x-go-package: github.com/MortalHappiness/VaccineReservationSystem/go-utils/models
host: localhost:7712
info:
  description: Documentation of Worker.
  title: Worker.
  version: '%VERSION_TO_BE_REPLACED_BY_GITRUNNER%'
paths:
  /.spec:
    get:
      description: GetSpec returns API Spec. This API is for testing & development
        only.
      operationId: GetSpec
      responses:
        "301":
          $ref: '#/responses/emptyResponse'
        "404":
          $ref: '#/responses/genericErrorcd'
      summary: Get Server API Specification
      tags:
      - Development
  /api/reservations/:reservationID:
    get:
      operationId: GetReservation
      responses:
        "200":
          $ref: '#/responses/ReservationResponse'
        "400":
          $ref: '#/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/responses/UnauthorizedErrorResponse'
        "404":
          $ref: '#/responses/NotFoundErrorResponse'
        "500":
          $ref: '#/responses/InternalServerErrorResponse'
      summary: Get the reservation information by reservation id.
      tags:
      - Reservation
  /api/reservations/users/:nationID:
    get:
      operationId: GetReservation
      responses:
        "200":
          $ref: '#/responses/ReservationResponse'
        "400":
          $ref: '#/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/responses/UnauthorizedErrorResponse'
        "404":
          $ref: '#/responses/NotFoundErrorResponse'
        "500":
          $ref: '#/responses/InternalServerErrorResponse'
      summary: Get the reservation information.
      tags:
      - Reservation
    post:
      operationId: PostReservationRequest
      parameters:
      - description: The reservation info
        in: body
        name: reservation
        schema:
          $ref: '#/definitions/ReservationModel'
        x-go-name: Reservation
      responses:
        "200":
          $ref: '#/responses/ReservationResponse'
        "400":
          $ref: '#/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/responses/UnauthorizedErrorResponse'
        "500":
          $ref: '#/responses/InternalServerErrorResponse'
      summary: Add a new reservation.
      tags:
      - Reservation
  /api/reservations/users/:nationID/:reservationID:
    delete:
      operationId: DeleteReservation
      responses:
        "200":
          $ref: '#/responses/DeleteReservationResponse'
          description: reservationID
        "400":
          $ref: '#/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/responses/UnauthorizedErrorResponse'
        "500":
          $ref: '#/responses/InternalServerErrorResponse'
      summary: Delete the reservation by id.
      tags:
      - Reservation
    put:
      operationId: PutReservationRequest
      parameters:
      - description: The reservation info
        in: body
        name: reservation
        schema:
          $ref: '#/definitions/ReservationModel'
        x-go-name: Reservation
      responses:
        "200":
          $ref: '#/responses/ReservationResponse'
        "400":
          $ref: '#/responses/BadRequestErrorResponse'
        "401":
          $ref: '#/responses/UnauthorizedErrorResponse'
        "500":
          $ref: '#/responses/InternalServerErrorResponse'
      summary: Add a new reservation.
      tags:
      - Reservation
produces:
- application/json
responses:
  BadRequestErrorResponse:
    description: BadRequestErrorResponse is the response of bad request error.
    schema:
      $ref: '#/definitions/APIError'
  InternalServerErrorResponse:
    description: InternalServerErrorResponse is the response of internal server error.
    schema:
      $ref: '#/definitions/APIError'
  NotFoundErrorResponse:
    description: NotFoundErrorResponse is the response of not found error.
    schema:
      $ref: '#/definitions/APIError'
  ReservationResponse:
    description: ReservationResponse is the response of GetReservation
    schema:
      $ref: '#/definitions/ReservationModel'
  UnauthorizedErrorResponse:
    description: UnauthorizedErrorResponse is the response of unauthorized error.
    schema:
      $ref: '#/definitions/APIError'
  emptyResponse:
    description: EmptyResponse are sent when the HTTP status code is 204.
  genericError:
    description: |-
      A GenericError is the default error message that is generated.
      For certain status codes there are more appropriate error structures.
    headers:
      message:
        description: |-
          The error message

          required: true
        type: string
schemes:
- http
swagger: "2.0"